requirements.txt
pip>=21.1.3
discord-py-slash-command>=2.3.2
